% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_insert_statements.R
\name{build_insert_statements}
\alias{build_insert_statements}
\title{Function to build \code{INSERT INTO} SQL statements from a data frame.}
\usage{
build_insert_statements(table, df, squish = FALSE)
}
\arguments{
\item{table}{Database table name to insert into.}

\item{df}{Data frame to use to create statements.}

\item{squish}{Should white space around commas and equal signs be removed to
make the SQL statement shorter?}
}
\value{
Character vector with a length of \code{nrow(df)}.
}
\description{
Function to build \code{INSERT INTO} SQL statements from a data frame.
}
\author{
Stuart K. Grange
}
