% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_update_statements.R
\name{build_update_statements}
\alias{build_update_statements}
\title{Function to build SQL \code{WHERE} statements from an input data frame.}
\usage{
build_update_statements(table, df, where = NA)
}
\arguments{
\item{table}{Database's table to update.}

\item{df}{Input data frame to generate \code{UPDATE} statements with.}

\item{where}{Which variable in \code{df} to be used for the \code{WHERE} 
clause. If not used, only a single row data frame can be used.}
}
\value{
Invisible.
}
\description{
Function to build SQL \code{WHERE} statements from an input data frame.
}
\author{
Stuart K. Grange
}
