% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.class.R
\name{db.class}
\alias{db.class}
\title{Function to determine database connection class/type.}
\usage{
db.class(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Character vector with a length of one.
}
\description{
\code{db.class} will return a lower-case string of database type which is 
suitable for logic testing similar to \code{\link{is.numeric}} or 
\code{\link{is.character}}.
}
\author{
Stuart K. Grange
}
