% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_analyse.R
\name{db_analyse}
\alias{db_analyse}
\title{Function to analyse/analyze a database table.}
\usage{
db_analyse(con, table)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table.}
}
\value{
Invisible.
}
\description{
Function to analyse/analyze a database table.
}
\author{
Stuart K. Grange
}
