% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_check_integrity.R
\name{db_check_integrity}
\alias{db_check_integrity}
\title{Function to test database integrity.}
\usage{
db_check_integrity(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Data frame.
}
\description{
\code{db_check_integrity} is only implemented for SQLite databases.
}
\author{
Stuart K. Grange
}
