% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection_name.R
\name{db_connection_name}
\alias{db_connection_name}
\title{Function to get a database connection string/name that is suitable for 
printing.}
\usage{
db_connection_name(con)
}
\arguments{
\item{con}{A database connection.}
}
\value{
A \strong{glue} character vector with length of \code{1}.
}
\description{
Function to get a database connection string/name that is suitable for 
printing.
}
\author{
Stuart K. Grange
}
