% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_copy_table_between_connections.R
\name{db_copy_table_between_connections}
\alias{db_copy_table_between_connections}
\title{Function to copy a database table from one connection to another.}
\usage{
db_copy_table_between_connections(
  con_source,
  con_destination,
  table,
  exists_only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{con_source}{Database connection to the source database.}

\item{con_destination}{Database connection to the destination database.}

\item{table}{Table to copy from \code{con_source} to \code{con_destination}.}

\item{exists_only}{Should only the tables which exist in both databases be
copied?}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{con_source}.
}
\description{
Function to copy a database table from one connection to another.
}
