% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_count_rows.R
\name{db_count_rows}
\alias{db_count_rows}
\title{Function to get row counts from database tables.}
\usage{
db_count_rows(con, table = NA, estimate = FALSE, progress = "none")
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table name in \code{con}. If \code{table} is unknown, all tables
in \code{con} will be queried.}

\item{estimate}{Only for PostgreSQL, should the table count estimate be used? 
Default is \code{FALSE}.}

\item{progress}{Type of progress bar to display. Default is \code{"none"}.}
}
\description{
The \code{row_count} variable will be a numeric value rather than an integer 
to avoid integer overflow issues for large tables.
}
\author{
Stuart K. Grange
}
