% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_count_variables.R
\name{db_count_variables}
\alias{db_count_variables}
\title{Function to count variables/columns/fields in a database table.}
\usage{
db_count_variables(con, table = NA)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table to get variables for. If \code{NA}, all tables will be 
queried.}
}
\description{
Function to count variables/columns/fields in a database table.
}
\author{
Stuart K. Grange
}
