% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_details.R
\name{db_details}
\alias{db_details}
\title{Function to return some high level details about a database.}
\usage{
db_details(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Tibble with one observation.
}
\description{
Function to return some high level details about a database.
}
\examples{
\dontrun{

# Get some information about the database
db_details(con)

}

}
\author{
Stuart K. Grange.
}
