% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_details.R
\name{db_details_insert}
\alias{db_details_insert}
\title{Function to return some high level details about a database.}
\usage{
db_details_insert(con, table = "database_details", print = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table to insert details into.}

\item{print}{Should \code{table} be printed post-insert?}
}
\value{
Invisible \code{con}.
}
\description{
Function to return some high level details about a database.
}
\examples{
\dontrun{

# Get some information about the database and insert into table
db_details_insert(con)

}

}
\author{
Stuart K. Grange.
}
