% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_direct_insert.R
\name{db_direct_insert}
\alias{db_direct_insert}
\title{Function to connect and insert directly into a database.}
\usage{
db_direct_insert(file, database = NA, table, df, replace = FALSE)
}
\arguments{
\item{file}{\code{JSON} file or string containing database connection 
details. For SQLite databases, use the database's file path. See 
\code{\link{db_connect}} for more information.}

\item{database}{The name of the database within \code{file} to use to create
a database connection to. If only one entry is in \code{file}, this argument 
is not needed and will be ignored if used.}

\item{table}{Table in, or to be created in \code{con}.}

\item{df}{Data frame to be inserted into \code{con}.}

\item{replace}{Should the database table be replaced? Default is \code{FALSE}. 
Be cautious using this argument because it will drop the database table if it
exists.}
}
\value{
Invisible.
}
\description{
\code{db_direct_insert} has been designed so a database can be connected to,
data inserted, and then disconnected with a single function call.
}
\seealso{
\code{\link{db_connect}}, \code{\link{db_insert}}, 
\code{\link{db_direct_read}}
}
\author{
Stuart K. Grange
}
