% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_direct_insert.R
\name{db_direct_list_tables}
\alias{db_direct_list_tables}
\title{Function to connect and list a database's tables directly.}
\usage{
db_direct_list_tables(file, database)
}
\arguments{
\item{file}{\code{JSON} file or string containing database connection 
details. For SQLite databases, use the database's file path. See 
\code{\link{db_connect}} for more information.}

\item{database}{The name of the database within \code{file} to use to create
a database connection to. If only one entry is in \code{file}, this argument 
is not needed and will be ignored if used.}
}
\value{
Character vector.
}
\description{
\code{db_direct_list_tables} has been designed so a database can be connected 
to, tables listed, data returned, and then disconnected with a single function
call.
}
\author{
Stuart K. Grange
}
