% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_direct_insert.R
\name{db_direct_read}
\alias{db_direct_read}
\title{Function to connect and read database table directly.}
\usage{
db_direct_read(file, database, table)
}
\arguments{
\item{file}{\code{JSON} file or string containing database connection 
details. For SQLite databases, use the database's file path. See 
\code{\link{db_connect}} for more information.}

\item{database}{The name of the database within \code{file} to use to create
a database connection to. If only one entry is in \code{file}, this argument 
is not needed and will be ignored if used.}

\item{table}{Table in, or to be created in \code{con}.}
}
\value{
Data frame.
}
\description{
\code{db_direct_read} has been designed so a database can be connected to,
data read, data returned, and then disconnected with a single function call.
}
\seealso{
\code{\link{db_connect}}, \code{\link{db_insert}}, 
\code{\link{db_direct_insert}}
}
\author{
Stuart K. Grange
}
