% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Function to close a database connection cleanly.}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{An active database connection.}
}
\value{
Invisible.
}
\description{
Function to close a database connection cleanly.
}
