% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_drop_table.R
\name{db_drop_table}
\alias{db_drop_table}
\title{Function to drop a table within a database.}
\usage{
db_drop_table(con, table)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table to drop.}
}
\value{
Invisible.
}
\description{
Function to drop a table within a database.
}
\author{
Stuart K. Grange.
}
