% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_duplicates.R
\name{db_duplicates}
\alias{db_duplicates}
\title{Function to find duplicates in a database table's variable.}
\usage{
db_duplicates(con, table, variable)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table.}

\item{variable}{Database table's variable to test.}
}
\value{
Data frame.
}
\description{
Function to find duplicates in a database table's variable.
}
\author{
Stuart K. Grange
}
