% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_execute.R
\name{db_execute}
\alias{db_execute}
\title{Function to execute statements on a database.}
\usage{
db_execute(con, statement, ..., progress = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{statement}{Statement to send to \code{con}.}

\item{...}{Other parameters passed on to methods.}

\item{progress}{Type of progress bar to display.}
}
\value{
Invisible \code{con}.
}
\description{
\code{db_execute} is a wrapper for \code{DBI::dbExecute} but is vectorised 
over \code{statement}.
}
\author{
Stuart K. Grange
}
