% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get.R
\name{db_get}
\alias{db_get}
\title{Function to get/fetch data from a database with a statement.}
\usage{
db_get(con, statement, warn = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{statement}{Statement to send to \code{con}.}

\item{warn}{Should the function return warnings?}
}
\value{
Tibble.
}
\description{
\code{db_get} is a wrapper for \code{DBI::dbGetQuery}.
}
\author{
Stuart K. Grange
}
