% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_duplicates.R
\name{db_has_duplicates}
\alias{db_has_duplicates}
\title{Function to test if duplicates are contained in a database table's variable.}
\usage{
db_has_duplicates(con, table, variable)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table.}

\item{variable}{Database table's variable to test.}
}
\value{
Logical vector with length of 1.
}
\description{
Function to test if duplicates are contained in a database table's variable.
}
\author{
Stuart K. Grange
}
