% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_information_schema.R
\name{db_information_schema}
\alias{db_information_schema}
\title{Function to get schema information from a database.}
\usage{
db_information_schema(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Only PostgreSQL databases are supported currently.
}
\author{
Stuart K. Grange
}

