% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_integrity_check.R
\name{db_integrity_check}
\alias{db_integrity_check}
\title{Function to test database integrity.}
\usage{
db_integrity_check(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Data frame.
}
\description{
\code{db_integrity_check} is only implemented for SQLite databases.
}
\author{
Stuart K. Grange
}
