% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_kill_process.R
\name{db_kill_process}
\alias{db_kill_process}
\title{Function to kill a database process.}
\usage{
db_kill_process(con, process)
}
\arguments{
\item{con}{Database connection.}

\item{process}{Process ID to kill. Use \code{\link{db_list_activities}} to 
find this integer.}
}
\description{
Function to kill a database process.
}
\author{
Stuart K. Grange
}
