% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_activities.R
\name{db_list_activities}
\alias{db_list_activities}
\title{Function to list activities/processes for a database.}
\usage{
db_list_activities(con, only_con = FALSE, as_json = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{only_con}{Filter activities only to the database where \code{con} has 
been made.}

\item{as_json}{Should the return be formatted as a JSON string?}
}
\description{
\code{db_list_activities} supports PostgreSQL and MySQL databases at present.
}
\author{
Stuart K. Grange
}
