% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_schemas.R
\name{db_list_schemas}
\alias{db_list_schemas}
\title{Function to list schemas for a database.}
\usage{
db_list_schemas(con, json = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{json}{Should the return be formatted as a JSON string.}
}
\description{
\code{db_list_schemas} only supports PostgreSQL databases at present.
}
\author{
Stuart K. Grange
}

