% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R
\name{db_list_tables}
\alias{db_list_tables}
\title{Function to list database tables.}
\usage{
db_list_tables(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Character vector.
}
\description{
Function to list database tables.
}
