% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_users.R
\name{db_list_users}
\alias{db_list_users}
\title{Function to list users for a database.}
\usage{
db_list_users(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Data frame.
}
\description{
Function to list users for a database.
}
\examples{
\dontrun{

# Get users
db_list_users(con)

}

}
\author{
Stuart K. Grange
}
