% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_name.R
\name{db_name}
\alias{db_name}
\title{Function to get database name.}
\usage{
db_name(con)
}
\arguments{
\item{con}{Database connection.}
}
\description{
Function to get database name.
}
\author{
Stuart K. Grange
}

