% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_post_gis_version.R
\name{db_post_gis_version}
\alias{db_post_gis_version}
\title{Function to return PostGIS version information from an PostGIS enabled 
PostgreSQL database.}
\usage{
db_post_gis_version(con, full = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{full}{Should full version information be returned?}
}
\value{
Character vector with the lenth of one.
}
\description{
Function to return PostGIS version information from an PostGIS enabled 
PostgreSQL database.
}
\author{
Stuart K. Grange
}
