% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R
\name{db_read_table}
\alias{db_read_table}
\title{Function to read an entire database table.}
\usage{
db_read_table(con, table, verbose = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table to read.}

\item{verbose}{Should the function print messages?}
}
\value{
Tibble.
}
\description{
\code{db_read_table} is a wrapper for \code{DBI::dbReadTable}.
}
