% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sqlite_master.R
\name{db_sqlite_master}
\alias{db_sqlite_master}
\title{Function to get SQLite \code{sqlite_master} query.}
\usage{
db_sqlite_master(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Data frame.
}
\description{
Function to get SQLite \code{sqlite_master} query.
}
\author{
Stuart K. Grange
}
