% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table_data_types.R
\name{db_table_data_types}
\alias{db_table_data_types}
\title{Function to get a database tables' data types.}
\usage{
db_table_data_types(con, table = NA)
}
\arguments{
\item{con}{Database connection.}

\item{table}{An optional character vector of table names to filter the return
to.}
}
\value{
Tibble.
}
\description{
Function to get a database tables' data types.
}
\author{
Stuart K. Grange
}
