% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R
\name{db_table_exists}
\alias{db_table_exists}
\title{Function to test if database contains a table.}
\usage{
db_table_exists(con, table)
}
\arguments{
\item{con}{Database connection.}

\item{table}{A table name.}
}
\value{
Logical vector.
}
\description{
Function to test if database contains a table.
}
