% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table_info.R
\name{db_table_info}
\alias{db_table_info}
\title{Function to get SQLite table information (\code{PRAGMA table_info}).}
\usage{
db_table_info(con, table = NA)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table name.}
}
\value{
Data frame.
}
\description{
Function to get SQLite table information (\code{PRAGMA table_info}).
}
\author{
Stuart K. Grange
}
