% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table_names.R
\name{db_table_names}
\alias{db_table_names}
\title{Function to get the names of database table and produce a data frame with 
zero rows.}
\usage{
db_table_names(con, table, as_tibble = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table.}

\item{as_tibble}{Should the return be a tibble?}
}
\value{
Data frame with zero rows.
}
\description{
\code{db_table_names} is useful when preparing a data frame for a SQL insert.
}
\author{
Stuart K. Grange
}
