% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table_permissions.R
\name{db_table_permissions}
\alias{db_table_permissions}
\title{Function to get a database's table permissions information for users and
tables.}
\usage{
db_table_permissions(con, user = NA, table = NA)
}
\arguments{
\item{con}{A database connection.}

\item{user}{A vector of database users.}

\item{table}{A vector of database tables.}
}
\value{
Tibble.
}
\description{
Currently, \code{db_table_permissions} only supports Postgres databases.
}
\author{
Stuart K. Grange
}
