% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_value_exists.R
\name{db_value_exists}
\alias{db_value_exists}
\title{Function to test if a value exists in a database table's column/variable.}
\usage{
db_value_exists(con, table, variable, value)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table in database.}

\item{variable}{Variable/column in \code{table}.}

\item{value}{Value to test in \code{table} and \code{variable}.}
}
\value{
Logical vector with length of 1.
}
\description{
Function to test if a value exists in a database table's column/variable.
}
\author{
Stuart K. Grange
}
