% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_truncate.R, R/db_version.R
\name{db_version}
\alias{db_version}
\alias{db_version}
\title{Function to run a \code{TRUNCATE TABLE} statement for a database table.}
\usage{
db_version(con)

db_version(con)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table to remove all data from.}

\item{con}{Database connection.}
}
\value{
Invisible.

Character vector
}
\description{
Beware that this function will remove the contents of a database table so use
with caution.

Function to get the version of a database service.
}
\author{
Stuart K. Grange

Stuart K. Grange
}
