% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_wildcard_check.R
\name{db_wildcard_check}
\alias{db_wildcard_check}
\title{Function to test for wildcards and then raise an error if found.}
\usage{
db_wildcard_check(x, type = "sql")
}
\arguments{
\item{x}{Argument to test.}

\item{type}{Type of wildcards to detect. Only SQL is implemented.}
}
\value{
Invisible \code{x}.
}
\description{
Function to test for wildcards and then raise an error if found.
}
\author{
Stuart K. Grange
}
