% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_formatter.R
\name{sql_formatter}
\alias{sql_formatter}
\title{Function to prepare an R data frame for SQL statement usage.}
\usage{
sql_formatter(df)
}
\arguments{
\item{df}{Data frame.}
}
\description{
\code{sql_formatter} will quote strings, factors, and dates and will also 
convert \code{NA}s to \code{NULL}s.
}
\author{
Stuart K. Grange
}
