% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_key_value.R
\name{sql_key_value}
\alias{sql_key_value}
\title{Function to create key-value strings for SQL statements.}
\usage{
sql_key_value(df)
}
\arguments{
\item{df}{Data frame which has been formated for SQL.}
}
\value{
List containing SQL key-value vectors.
}
\description{
Function to create key-value strings for SQL statements.
}
\author{
Stuart K. Grange
}
\seealso{
\code{\link{sql_formatter}}
}

