% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sql_collapse.R
\name{str_sql_collapse}
\alias{str_sql_collapse}
\title{Function to collapse vectors into a vector with a length of one for use in
SQL statements.}
\usage{
str_sql_collapse(x)
}
\arguments{
\item{x}{A vector to be collapsed.}
}
\value{
A character vector with a length of one.
}
\description{
\code{str_sql_collapse} will quote character and factor vectors before being
collapsed.
}
\examples{

# Collapse an integer vector
str_sql_collapse(1:10)

# Collapse a character vector
str_sql_collapse(letters[1:10])

}
\author{
Stuart K. Grange
}
