% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_update_statements.R
\name{build_update_statements}
\alias{build_update_statements}
\title{Function to build SQL \code{WHERE} statements from an input data frame.}
\usage{
build_update_statements(table, df, where = NA, squish = FALSE)
}
\arguments{
\item{table}{Database's table to update.}

\item{df}{Input data frame to generate \code{UPDATE} statements with.}

\item{where}{Which variables in \code{df} to be used for the \code{WHERE} 
clause. If not used, only a single row data frame can be used.}

\item{squish}{Should whitespace around commas and equal signs be removed to
make the SQL statement shorter?}
}
\value{
Character vector with a length of \code{nrow(df)}.
}
\description{
Function to build SQL \code{WHERE} statements from an input data frame.
}
\author{
Stuart K. Grange
}
