% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection_name.R
\name{db_connection_name}
\alias{db_connection_name}
\alias{cli_db_connection}
\title{Functions to get a database connection strings/names that are suitable for 
printing.}
\usage{
db_connection_name(con)

cli_db_connection(con)
}
\arguments{
\item{con}{A database connection.}
}
\value{
A character or \strong{glue} character vector with length of \code{1}.
}
\description{
Functions to get a database connection strings/names that are suitable for 
printing.
}
\author{
Stuart K. Grange
}
