% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_count_variables.R
\name{db_count_variables}
\alias{db_count_variables}
\title{Function to count variables/columns/fields in a database table.}
\usage{
db_count_variables(con, table = NA)
}
\description{
Function to count variables/columns/fields in a database table.
}
\author{
Stuart K. Grange
}
