% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_data_types.R
\name{db_list_data_types}
\alias{db_list_data_types}
\title{Function to list data types of database tables.}
\usage{
db_list_data_types(con, table = NA, json = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Tables to get data types for. If not used, all tables will be
returned.}

\item{Should}{the return be in pretty JSON? Default is \code{FALSE}.}
}
\description{
\code{db_list_data_types} only supports PostgreSQL databases at present.
}
\author{
Stuart K. Grange
}
