% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_list_indices.R
\name{db_list_indices}
\alias{db_list_indices}
\title{Function to list indices for a database tables.}
\usage{
db_list_indices(con)
}
\arguments{
\item{con}{Database connection.}

\item{table}{A vector of table names. If unused, all tables will be queried.}
}
\description{
Function to list indices for a database tables.
}
\author{
Stuart K. Grange
}
