% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_simple_wrappers.R,
%   R/dplyr_re_database_exports.R
\docType{import}
\name{db_list_tables}
\alias{db_list_tables}
\alias{reexports}
\alias{db_list_tables}
\alias{reexports}
\alias{db_has_table}
\title{Function to list database tables.}
\usage{
db_list_tables(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Character vector.
}
\description{
Function to list database tables.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{db_list_tables}}, \code{\link[dplyr]{db_has_table}}}
}}

