% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_name.R
\name{db_name}
\alias{db_name}
\title{Function to get database name.}
\usage{
db_name(con, extension = TRUE)
}
\arguments{
\item{con}{Database connection.}

\item{extension}{For SQLite databases, should the database name include the
file name extension?}
}
\description{
Function to get database name.
}
\author{
Stuart K. Grange
}
