% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_size.R
\name{db_size}
\alias{db_size}
\title{Function to get the size of a database.}
\usage{
db_size(con, unit = "mb")
}
\arguments{
\item{con}{Database connection.}

\item{unit}{Measurement unit. Default is \code{"mb"} for megabytes.}
}
\value{
Numeric vector.
}
\description{
For SQLite databases, the file size is returned, for PostgreSQL databases,
the `pg_database` table is queried, and for MySQL database 
`information_schema.tables` is queried.
}
\author{
Stuart K. Grange
}
