% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_truncate_table.R
\name{db_truncate_table}
\alias{db_truncate_table}
\title{Function to run a \code{TRUNCATE TABLE} statement for a database table.}
\usage{
db_truncate_table(
  con,
  table,
  cascade = FALSE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Table to remove all data from.}

\item{cascade}{Should the \code{TRUNCATE} statement be used with 
\code{CASCADE}? Using \code{cascade} will delete relations which are 
dependent on \code{table}.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Invisible \code{con}.
}
\description{
Beware that this function will remove the contents of a database table, and
possibly other tables so use with caution.
}
\author{
Stuart K. Grange
}
