% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_vacuum.R
\name{db_vacuum}
\alias{db_vacuum}
\alias{db_optimise}
\title{Functions to vacuum, analyse, and optimise database tables.}
\usage{
db_vacuum(con, table, return = FALSE)

db_optimise(con, table, return = FALSE)
}
\arguments{
\item{con}{Database connection.}

\item{table}{Database table.}

\item{return}{Should the function return a data frame containing the database
sizes pre- and post-vacuum?}
}
\value{
Invisible or data frame if \code{return = TRUE}.
}
\description{
The PostgreSQL statement uses \code{VACUUM (VERBOSE)}, the MySQL statement
uses \code{OPTIMIZE}, and the SQLite statement uses \code{VACUUM} and the 
\code{table} argument is not used.
}
\examples{
\dontrun{

db_vacuum_analyse(con, "model_data")
db_optimise(con, "air_quality_data")

}

}
\author{
Stuart K. Grange
}
