% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_version.R
\name{db_version}
\alias{db_version}
\title{Function to get the version of a database service.}
\usage{
db_version(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Character vector
}
\description{
Function to get the version of a database service.
}
\author{
Stuart K. Grange
}
