% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_re_exports_with_dbplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{db_list_tables}
\alias{reexports}
\alias{db_has_table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{db_list_tables}}, \code{\link[dplyr]{db_has_table}}}
}}

