% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update_builder.R
\name{sql_update_builder}
\alias{sql_update_builder}
\title{Function to build SQL UPDATE statements.}
\usage{
sql_update_builder(sql, table)
}
\value{
Character vector.
}
\description{
\code{sql_update_builder} will use the first key-value pair to build the 
WHERE clause. To-do: make better.
}
\author{
Stuart K. Grange
}
